Fix vehicle coin saving.
The previous method for saving coins only worked for static class names
and is checked everytime server_updateObject is referenced so it's quite
intensive.
Since there are too many vehicles to list, we'll just make a proper
server_updateObject function specific to saving coins on objects which
will be a lot less overhead.